+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Enable gradient accumulation using Horovod's `backward_passes_per_step`
+
+
 - Add new `DETAIL` log level to provide useful logs for improving monitoring and debugging of batch jobs